refactor: 당첨 결과 매핑 역할을 Lottos로 이동
