+++ src/main/java/lotto/domain/Lottos.java
@@ -1,7 +1,9 @@
 package lotto.domain;
 
 import java.util.ArrayList;
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 import lotto.domain.lottogenerator.LottoGenerator;
 
 public class Lottos {
@@ -22,6 +24,34 @@ public class Lottos {
 
     public static Lottos withListLotto(List<Lotto> lottos) {
         return new Lottos(lottos);
+    }
+
+    public Map<Rank, Integer> mapResult(WinningLotto winningLotto) {
+        Map<Rank, Integer> result = initResult();
+
+        for (Lotto lotto : lottos) {
+            Rank rank = getRank(winningLotto, lotto);
+            if (rank == Rank.NONE) continue;
+            result.put(rank, result.get(rank) + 1);
+        }
+
+        return result;
+    }
+
+    private Map<Rank, Integer> initResult() {
+        Map<Rank, Integer> result = new LinkedHashMap<>();
+        for (Rank rank : Rank.values()) {
+            if (rank == Rank.NONE) continue;
+            result.put(rank, 0);
+        }
+
+        return result;
+    }
+
+    private Rank getRank(WinningLotto winningLotto, Lotto userLotto) {
+        int matchCount = winningLotto.getMatchCount(userLotto);
+        boolean matchBonus = winningLotto.getMatchBonus(userLotto);
+        return Rank.find(matchCount, matchBonus);
     }
 
     public List<Lotto> get() {
